# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for GMarket",
        "title": "GMarket API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/market/chart": {
            "get": {
                "description": "GetChartData fetches chart data for a given symbol from an external API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "market"
                ],
                "summary": "Get Chart Data for a Symbol",
                "parameters": [
                    {
                        "type": "string",
                        "default": "NSE",
                        "description": "Exchange (e.g., NSE)",
                        "name": "exchange",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "RELIANCE",
                        "description": "Stock Symbol (e.g., RELIANCE)",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1d",
                        "description": "Chart Range (e.g., 1d, 5d, 1mo)",
                        "name": "range",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1m",
                        "description": "Data Interval (e.g., 1m, 5m, 1h)",
                        "name": "interval",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/market/symbols": {
            "get": {
                "description": "GetSymbols serves the list of market symbols from a static JSON file.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "market"
                ],
                "summary": "Get Market Symbols",
                "responses": {
                    "200": {
                        "description": "../data/stocks.json",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/sessions": {
            "post": {
                "description": "Create a new session for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Create a new session (login)",
                "parameters": [
                    {
                        "description": "User Login Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateSessionDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.SessionDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the current session for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Delete the current session (logout)",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Retrieve information about the currently authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user",
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User Registration Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UserRegistrationDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the currently authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete current user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/watchlists/{watchlist_idx}": {
            "get": {
                "description": "Retrieve the specified watchlist for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlists"
                ],
                "summary": "Get a user's watchlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Watchlist Index (1-10)",
                        "name": "watchlist_idx",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.WatchlistDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/watchlists/{watchlist_idx}/symbols": {
            "post": {
                "description": "Add a stock symbol to the specified watchlist for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlists"
                ],
                "summary": "Add a symbol to a user's watchlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Watchlist Index (1-10)",
                        "name": "watchlist_idx",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Symbol to add",
                        "name": "symbol",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.SymbolDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a stock symbol from the specified watchlist for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlists"
                ],
                "summary": "Remove a symbol from a user's watchlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Watchlist Index (1-10)",
                        "name": "watchlist_idx",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Symbol to remove",
                        "name": "symbol",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.SymbolDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.CreateSessionDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "strongpassword123"
                }
            }
        },
        "dtos.ErrorDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "dev_message": {
                    "type": "string",
                    "example": "Invalid request body format"
                },
                "message": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "dtos.SessionDTO": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/dtos.UserDTO"
                }
            }
        },
        "dtos.SymbolDTO": {
            "type": "object",
            "properties": {
                "exchange": {
                    "type": "string",
                    "example": "NSE"
                },
                "symbol": {
                    "type": "string",
                    "example": "RELIANCE"
                }
            }
        },
        "dtos.SymbolInfo": {
            "type": "object",
            "properties": {
                "exchange": {
                    "type": "string",
                    "example": "NSE"
                },
                "last_close_price": {
                    "type": "number",
                    "example": 2430.5
                },
                "ltp": {
                    "type": "number",
                    "example": 2450.75
                },
                "symbol": {
                    "type": "string",
                    "example": "RELIANCE"
                }
            }
        },
        "dtos.UserDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-10-01T12:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "profile_picture_url": {
                    "type": "string",
                    "example": "https://example.com/profile.jpg"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-10-01T12:00:00Z"
                }
            }
        },
        "dtos.UserRegistrationDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "password": {
                    "type": "string",
                    "example": "strongpassword123"
                }
            }
        },
        "dtos.WatchlistDTO": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer",
                    "example": 1
                },
                "symbols": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.SymbolInfo"
                    }
                }
            }
        }
    }
}